backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
httpx==0.27.0
beautifulsoup4==4.12.3
PyPDF2==3.0.1
python-multipart==0.0.9
boto3==1.34.149
python-dotenv==1.0.1
pydantic==2.8.2
motor==3.5.1
pymongo==4.8.0
playwright==1.45.0
lxml==5.2.2
html2text==2024.2.26